Internal: move Cairo 0 hash-related hints to dedicated module
This makes it easier to identify changes related to hashing.
